Cleaned the code(w sumie to nasrane jeszcze bardziej)
