+++ src/app/app-routing.module.ts
@@ -3,8 +3,8 @@ import { Routes, RouterModule } from '@angular/router';
 import { DashboardComponent } from './components/dashboard/dashboard.component';
 import { HomePageComponent } from './components/home-page/home-page.component';
 import { LoginComponent } from './components/login/login.component';
-import { RegisterMockComponent } from './components/register-mock/register-mock.component';
-import { RegisterComponent } from './components/register/register.component';
+import { RegisterMockComponent } from './components/register/register.component';
+import { RegisterComponent } from './components/register-unused/register.component';
 import { StatisticsComponent } from './components/statistics/statistics.component';
 import { AuthGuard } from './_helpers';
 
@@ -14,11 +14,10 @@ const usersModule = () => import('@app/components/childs/users.module').then(x =
 const routes: Routes = [
 { path: '', redirectTo: '/home', pathMatch: 'full' },
 { path: 'home', component: HomePageComponent },
-{ path: 'register', component: RegisterComponent },
 { path: 'dashboard', component: DashboardComponent, canActivate: [AuthGuard] },
 { path: 'login', component: LoginComponent },
-{ path: 'register-mock', component: RegisterMockComponent },
-{ path: 'users', loadChildren: usersModule, canActivate: [AuthGuard] },
+{ path: 'register', component: RegisterMockComponent },
+{ path: 'children', loadChildren: usersModule, canActivate: [AuthGuard] },
 { path: 'statistics', component: StatisticsComponent }
 ];
