+++ src/app/app.component.html
@@ -2,9 +2,9 @@
     <nav class="nav">
       <ul>
         <li *ngIf="!user" class="nav-primary"><a routerLink="/home">Strona Główna</a></li>
-        <li *ngIf="!user" class="nav-primary"><a routerLink="/register-mock">Rejestracja</a></li>
+        <li *ngIf="!user" class="nav-primary"><a routerLink="/register">Rejestracja</a></li>
         <li *ngIf="!user" class="nav-primary"><a routerLink="/login">Logowanie</a></li>
-        <li *ngIf="user" class="nav-primary"><a routerLink="/users" routerLinkActive>Twoje dzieci</a></li>
+        <li *ngIf="user" class="nav-primary"><a routerLink="/children" routerLinkActive>Twoje dzieci</a></li>
         <li *ngIf="user" class="nav-primary"><a routerLink="/statistics" >Statystyki</a></li>
         <li *ngIf="user" class="nav-primary"><a routerLink="">NIEZBLIŻSIĘTU</a></li>
         <li *ngIf="user" class="nav-primary"><a (click)="logout()">Wyloguj</a></li>
