+++ src/app/components/childs/add-edit.component.html
@@ -59,6 +59,6 @@
             <span *ngIf="loading" class="spinner-border spinner-border-sm mr-1"></span>
             Save
         </button>
-        <a routerLink="/users" class="btn btn-link">Cancel</a>
+        <a routerLink="/children" class="btn btn-link">Cancel</a>
     </div>
 </form>
