--- src/app/components/register-mock/register-mock.component.html
@@ -1,53 +1,0 @@
-<div class="l-nav">
-    <nav class="nav">
-      <ul>
-        <li class="nav-primary"><a routerLink="/home">Strona Główna</a></li>
-        <li class="nav-primary"><a routerLink="/register-mock">Rejestracja</a></li>
-        <li class="nav-primary"><a routerLink="/login">Logowanie</a></li>
-      </ul>
-    </nav>
-  </div>     
-
-<div class="card">
-    <h4 class="card-header">Register</h4>
-    <div class="card-body">
-        <form [formGroup]="form" (ngSubmit)="onSubmit()">
-            <div class="form-group">
-                <label for="username">Username</label>
-                <input type="text" formControlName="username" class="form-control" [ngClass]="{ 'is-invalid': submitted && f.username.errors }" />
-                <div *ngIf="submitted && f.username.errors" class="invalid-feedback">
-                    <div *ngIf="f.username.errors.required">Username is required</div>
-                </div>
-            </div>    
-            <div class="form-group">
-                <label for="birthdate">Birthdate</label>
-                <input type="date" max="{{isAdult}}" formControlName="birthdate" class="form-control" [ngClass]="{ 'is-invalid': submitted && f.birthdate.errors }" />
-                <div *ngIf="submitted && f.birthdate.errors" class="invalid-feedback">
-                    <div *ngIf="f.birthdate.errors.required">Birthdate is required</div>
-                </div>
-            </div>   
-            <div class="form-group">
-                <label for="email">email</label>
-                <input type="email" formControlName="email" class="form-control" [ngClass]="{ 'is-invalid': submitted && f.email.errors }" />
-                <div *ngIf="submitted && f.email.errors" class="invalid-feedback">
-                    <div *ngIf="f.email.errors.required">email is required</div>
-                </div>
-            </div>                   
-            <div class="form-group">
-                <label for="password">Password</label>
-                <input type="password" formControlName="password" class="form-control" [ngClass]="{ 'is-invalid': submitted && f.password.errors }" />
-                <div *ngIf="submitted && f.password.errors" class="invalid-feedback">
-                    <div *ngIf="f.password.errors.required">Password is required</div>
-                    <div *ngIf="f.password.errors.minlength">Password must be at least 6 characters</div>
-                </div>
-            </div>
-            <div class="form-group">
-                <button [disabled]="loading" class="btn btn-primary">
-                    <span *ngIf="loading" class="spinner-border spinner-border-sm mr-1"></span>
-                    Register
-                </button>
-                <a routerLink="../login" class="btn btn-link">Cancel</a>
-            </div>
-        </form>
-    </div>
-</div>
