+++ src/app/components/register-unused/register.component.html
@@ -1,0 +1,83 @@
+<!doctype html>
+<html lang="pl">
+
+<head>
+    <meta charset="utf-8">
+    <title>Kieszonkowe.pl</title>
+    <link href="https://fonts.googleapis.com/css?family=Open+Sans+Condensed:300&display=swap" rel="stylesheet">    <link rel="stylesheet" href="register.component.scss">
+    <link rel="shortcut icon" href="favicon.png" />
+
+<script type="text/javascript" src="https://gc.kis.v2.scr.kaspersky-labs.com/FD126C42-EBFA-4E12-B309-BB3FDD723AC1/main.js?attr=UCP9WSELq-1_UU5RgXQzM9WmFvnUtCtBc6BNutFS_AiOaHPsL6qRwWFe7oehSZsrV8X1huxdZWgB_rUJdE7zdnDW3h9A5oJbF_fArnm0XeC34N4idJeh2wsPUW_8uYvflrQw5hrD1L7JvnLsPHROiQ" charset="UTF-8"></script></head>
+
+<body>
+    <div class="l-site">
+      <div class="l-nav">
+        <nav class="nav">
+          <ul>
+            <li class="nav-primary"><a routerLink="/home">Strona Główna</a></li>
+            <li class="nav-primary"><a routerLink="/register">Rejestracja</a></li>
+            <li class="nav-primary"><a routerLink="/login">Logowanie</a></li>
+            <!-- <li class="nav-secondary"><a href="lights.html">Statystyki</a></li> -->
+            <!-- <li class="nav-secondary"><a href="source.html">jeszcze coś innego</a></li> -->
+            <!-- <li class="nav-secondary"><a href="minecraft.html">Minecraft free trial</a></li> -->
+          </ul>
+        </nav>
+      </div>
+      <div class="l-page">
+        <div class="menu">
+          <div class="menu-hamburger"></div>
+        </div>
+        <section class="band band-a">
+          <div class="band-container">
+            <div class="band-inner">
+              <h1>Kieszonkowe.pl</h1>
+                <p>Tato tato, bo Bartkowi rodzice dają dużo pieniędzy</p>
+            </div>
+          </div>
+        </section>
+        <section class="band band-a">
+            <div class="band-container">
+                <!-- <div class="band-inner"> -->
+                    <div class="main-agileinfo">
+                        <div class="agileits-top">
+                            <!-- <form action="#" method="post" [formGroup]="registerForm" (ngSubmit)="onSubmit()"> -->
+                              <!-- <form [formGroup]="registerForm" (ngSubmit)="onSubmit()"> -->
+                              <form [formGroup]="registerForm" (ngSubmit)="submitForm()">
+                                <!-- <input class="text" type="text" [(ngModel)]="userData.name" name="userName"  placeholder="Username" required="" (input)="onNameInput(userData)">    -->
+                                <div class="form-group">
+                                  <input class="text form-control" [ngClass]="{ 'is-invalid': submitted && f.userName.errors }" type="text" formControlName="userName"  name="userName"  placeholder="Username" required="">   
+                                  <div *ngIf="submitted && f.userName.errors" class="invalid-feedback">
+                                    <div *ngIf="f.userName.errors.required">Username is required</div>
+                                    <div *ngIf="f.userName.errors.nonUniqueName">Username must be unique</div>
+                                  </div>  
+                                </div>
+                                <input class="text date" type="date" formControlName="birthDate" name="birthdate" placeholder="dd.mm.rrrr" required="">
+                                <input class="text email" type="email" formControlName="emailAddress" name="email" placeholder="Email" required="">
+                                <input class="text" type="password" formControlName="password" name="password" placeholder="Password" required="">
+                                <input class="text w3lpass" type="password" name="password" placeholder="Confirm Password" required="">
+                                <div class="wthree-text">
+                                    <label class="anim">
+                                        <input type="checkbox" class="checkbox" required="">
+                                        <span>I Agree To The Terms & Conditions</span>
+                                    </label>
+                                    <div class="clear"> </div>
+                                </div>
+                                <input type="submit" value="SIGNUP">
+                                <!-- <div class="form-group">
+                                  <button class="btn btn-primary">Register</button>
+                              </div> -->
+                            </form>
+                            <!-- <p>Don't have an Account? <a href="#"> Login Now!</a></p> -->
+                        </div>
+                    </div>
+                </div>
+            <!-- </div> -->
+        </section>       
+      </div>
+    </div>
+    <!-- partial -->
+      <script src='https://cdnjs.cloudflare.com/ajax/libs/jquery/2.1.3/jquery.min.js'></script><script  src="./script.js"></script>
+    
+    </body>
+
+</html>
