+++ src/app/components/register-unused/register.component.scss
@@ -7,9 +7,6 @@ body {
   color: rgb(228, 206, 15);
   font-family: "Open Sans", sans-serif;  
 }
-.card {
-    margin-left: 400px;
-}
 
 .invalid-feedback {
   color: red;
@@ -404,7 +401,7 @@ License URL: http://creativecommons.org/licenses/by/3.0/
 
 input[type="text"], input[type="email"], input[type="password"], input[type="date"] {
   font-size: 0.9em;
-  color: rgb(73, 209, 243);
+  color: #fff;
   font-weight: 100;
   width: 95%;
   display: block;
@@ -417,7 +414,6 @@ input[type="text"], input[type="email"], input[type="password"], input[type="dat
   background: linear-gradient(to bottom, rgba(255, 255, 255, 0) 96%, #fff 4%);
   background-position: -800px 0;
   background-size: 100%;
-  background-color: rgb(30, 78, 3);
   background-repeat: no-repeat;
   color: #fff;
   font-family: 'Roboto', sans-serif;
