+++ src/app/components/register/layout.component.ts
