+++ src/app/components/statistics/statistics.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { StatisticsService } from 'app/services/statistics.service';
 
+import { first } from 'rxjs/operators';
+
 @Component({
   selector: 'app-statistics',
   templateUrl: './statistics.component.html',
@@ -17,6 +19,10 @@ export class StatisticsComponent implements OnInit {
 
   ngOnInit(): void {
     this.statisticsService.getEducations().subscribe((educations) => (this.educations = educations));
+    // this.statisticsService.getEducations()
+    // .pipe(first())
+    // .subscribe((educations) => (this.educations = educations));
+
     this.statisticsService.getRegions().subscribe((regions) => (this.regions = regions));
   }
 }
