+++ src/app/services/personal-data.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
-import { PersonalData } from  '@app/components/register/personalData';
-import { personData } from '@app/components/register/person-mock';
+import { PersonalData } from  '@app/components/register-unused/personalData';
+import { personData } from '@app/components/register-unused/person-mock';
 
 @Injectable({
   providedIn: 'root'
